Implemented Product adding and getting all
Generalized product adding, wrapped the ObejctMapper into a Factory, tried to keep it as generalized and simple as possible.

Added multiple enums,might make it more in the future, also we will only use the objectMapper at adding a product, next for the update request or responses DTO-s will be introduced.